move PipelineRunner and PipelineTracker into their own files
