+++ tests/test_workflow_pipeline.py
@@ -1,4 +1,4 @@
-"""Tests for _PipelineWorkflow, PipelineRunner, and PipelineTracker."""
+"""Tests for BasePipelineWorkflow."""
 
 import json
 from logging import Logger
@@ -11,10 +11,10 @@ from fids import fids
 from nipoppy.config.base import Config, PipelineConfig
 from nipoppy.config.boutiques import BoutiquesConfig
 from nipoppy.utils import strip_session
-from nipoppy.workflows.pipeline import PipelineRunner, _PipelineWorkflow
+from nipoppy.workflows.pipeline import BasePipelineWorkflow
 
 
-class PipelineWorkflow(_PipelineWorkflow):
+class PipelineWorkflow(BasePipelineWorkflow):
     def __init__(
         self,
         dpath_root: Path | str,
@@ -101,7 +101,7 @@ class PipelineWorkflow(_PipelineWorkflow):
 )
 def test_init(args):
     workflow = PipelineWorkflow(**args)
-    assert isinstance(workflow, _PipelineWorkflow)
+    assert isinstance(workflow, BasePipelineWorkflow)
     assert hasattr(workflow, "pipeline_name")
     assert hasattr(workflow, "pipeline_version")
     assert hasattr(workflow, "participant")
@@ -402,73 +402,3 @@ def test_generate_fpath_log(
     )
     fpath_log = workflow.generate_fpath_log()
     assert fpath_log.stem.startswith(expected_stem)
-
-
-@pytest.mark.parametrize("simulate", [True, False])
-def test_runner(simulate, tmp_path: Path):
-    pipeline_name = "dummy_pipeline"
-    pipeline_version = "1.0.0"
-    runner = PipelineRunner(
-        dpath_root=tmp_path / "my_dataset",
-        pipeline_name=pipeline_name,
-        pipeline_version=pipeline_version,
-        simulate=simulate,
-    )
-    runner.config = Config(
-        DATASET_NAME="my_dataset",
-        SESSIONS=["ses-BL", "ses-V04"],
-        SINGULARITY_CONFIG={"COMMAND": "echo"},  # dummy command
-        PROC_PIPELINES={
-            pipeline_name: {
-                pipeline_version: {
-                    "DESCRIPTOR": {
-                        "name": pipeline_name,
-                        "tool-version": pipeline_version,
-                        "description": "A dummy pipeline for testing",
-                        "schema-version": "0.5",
-                        "command-line": "echo [ARG1] [ARG2] [[NIPOPPY_DPATH_BIDS]]",
-                        "inputs": [
-                            {
-                                "id": "arg1",
-                                "name": "arg1",
-                                "type": "String",
-                                "command-line-flag": "--arg1",
-                                "value-key": "[ARG1]",
-                            },
-                            {
-                                "id": "arg2",
-                                "name": "arg2",
-                                "type": "Number",
-                                "command-line-flag": "--arg2",
-                                "value-key": "[ARG2]",
-                            },
-                        ],
-                    },
-                    "INVOCATION": {
-                        "arg1": "[[NIPOPPY_PARTICIPANT]] [[NIPOPPY_SESSION]]",
-                        "arg2": 10,
-                    },
-                    "SINGULARITY_CONFIG": {
-                        "COMMAND": "echo",
-                    },
-                }
-            }
-        },
-    )
-
-    participant = "01"
-    session = "ses-BL"
-
-    fids.create_fake_bids_dataset(
-        runner.layout.dpath_bids,
-        subjects=participant,
-        sessions=strip_session(session),
-    )
-
-    runner.dpath_pipeline_output.mkdir(parents=True, exist_ok=True)
-    runner.dpath_pipeline_work.mkdir(parents=True, exist_ok=True)
-    descriptor_str, invocation_str = runner.run_single(participant, session)
-
-    assert "[[NIPOPPY_DPATH_BIDS]]" not in descriptor_str
-    assert "[[NIPOPPY_PARTICIPANT]]" not in invocation_str
-    assert "[[NIPOPPY_SESSION]]" not in invocation_str
